+++ app/src/main/java/io/gnosis/safe/ui/safe/terms/TermsViewModel.kt
@@ -28,8 +28,4 @@ class TermsViewModel @Inject constructor(
         termsChecker.setTermsAgreed(true)
         advance()
     }
-
-    companion object {
-        const val TERMS_AGREED = "prefs.boolean.terms_agreed"
-    }
 }
